Refactor test_reddit.py to reddit_test.py
